Add ability to sort reviews (#23)
* Update eslintrc to warn for index keys

Using index keys is non-optimal, but is sometimes the only valid option
(other than not setting a key, which implicitly does the same thing).  I
would rather do it explicitly and get a warning—and thus a reminder that
it isn't optimal—than do it implicitly.

* Add RatingSnapshot component

This component presents an overview of the total ratings for an item.

This component still needs tests

* Add AverageRating component

* Add tests for AverageRating and RatingSnapshot

* Add SortSelector component

* Add tests for initial sorting

* Add ability to sort by date

* Fix issue with date sorting

Ensures that dates sort correctly.  Sorts by date on initial page load

* Add date column in package.json

* Fix sorting tests

* Add jest-coveralls support to Travis

* Add test of basic page load

* Fix broken link in README