Increment helpful button (#26)
* Refactor code to break up long functions

CodeClimate detected three functions that exceeded 25 substantive lines
of code.  In all three cases, this was because the functions included
styled-components styles inside the function.  I've refactored all three
to move the style above the function, which better separates the styling
from the logic/structure of the component.

* Refactor to use props restructuring in argument

This is a shift from

```
props => {
  const { variableOfInterest }
}

```
to

```
({ variableOfInterest }) => {}
```

Which means the same thing but is shorter/more readable.

* Update merge dependency to address security

* Add tests for incrementing helpfulness data

* Add incrementHelpful method

* Add ability to change helpfulness data on frontend

* Update back end to persist helpfulness changes